Split out config.py
