+++ configs/10B/H100_devel.toml
@@ -1,0 +1,23 @@
+name_model = "10B"
+project = "debug_I2_zero_band"
+
+[train]
+micro_bs = 1
+ac_ckpt = true
+
+[optim]
+sched_type = "wsd-sqrt"
+batch_size = 128 #1M tokens bs
+warmup_steps = 0
+total_steps = 1
+lr = 7.5e-5
+
+adam_betas1 = 0.9
+adam_betas2 = 0.95
+weight_decay = 0.1
+
+z_loss = true
+
+[data]
+seq_length = 8192
+num_workers = 4
