+++ src/zeroband/data.py
@@ -416,6 +416,9 @@ def load_all_datasets(
         split_rank = rank
         split_world_size = world_size
 
+
+    logger.info(f"Loading Train dataset(s)")
+
     ds = _load_datasets(
         dataset_names=data_config.dataset_name_or_paths,
         split=split,
